requirements.txt
python-docx>=0.8.11
Pillow>=9.0.0
configparser>=5.0.0